requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
APScheduler==3.10.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.2
betterproto==2.0.0b6
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
decorator==5.1.1
fonttools==4.41.0
frozenlist==1.4.0
gevent==23.7.0
gevent-eventemitter==2.1
greenlet==2.0.2
grpclib==0.4.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
idna==3.4
kiwisolver==1.4.4
lxml==4.9.3
matplotlib==3.7.2
multidict==6.0.4
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
protobuf==3.20.3
pyaes==1.6.1
pycparser==2.21
pycryptodomex==3.18.0
pyparsing==3.0.9
Pyrogram==2.0.106
pyropatch==2.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-toolbelt==1.0.0
schedule==1.2.0
seaborn==0.12.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
steamio==1.0.1
telegraph==2.2.0
TgCrypto==1.2.5
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
vdf==3.4
yarl==1.9.2
zope.event==5.0
zope.interface==6.0
